Moves EarlyAccessFlag to separate file
